Update patch set 1
Patch Set 1:

My main concern about implementing this, is that if you're running with multiple threads, one thread does its Final Select and exits, if the other thread is still running and updating the database, the first thread will be out of sync anyway.  Not that this is wrong, but it may end up being incomplete.

Patch-set: 1
Label: Verified=0